requirements.txt
Flask==1.1.2
pydantic==1.8.1
google-cloud-datastore==2.1.2
